Revert "Split zsync_begin into two parts, allowing for zsync_rename earlier."
... and followup commites commits

The original design was that one would first call zsync_begin,
then do the transfer, and when finished, do zsync_rename in order
to rename the tmp file to the final file.

But the changes being reverted made it so that one first cakk zsync_parse,
(and this creates an empty the temporary file in the current directory),
then one call zsync_rename to rename it to the name of the temporary file that owncloud wants.
But zsync_rename no longer rename the file (just remove the old one)
and this is a change of semantic that breaks the zsync client.

Instead, I'll change the API so zsync_begin will contains the
filename of the temportary file

This reverts commit ae5ea93593e218646b0cd266841711ee55449b48,
commit 85b25da0c5ac82b42eabed8e47ba2728ddbb62ff, commit
3271b60504916c614448acf00ab3437a897aa1df, and part of commit
fcf19315a565c18e34e7c5ddbf23cb0dd6846491.

 Conflicts:
	c/librcksum/state.c
	c/libzsync/zsync.c